requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
executing==2.2.0
fastapi==0.103.1
frozenlist==1.6.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain-core==0.3.63
langchain-openai==0.3.19
langsmith==0.3.44
matplotlib-inline==0.1.7
multidict==6.4.4
nest-asyncio==1.6.0
numpy==1.24.3
openai==1.84.0
orjson==3.10.18
packaging==24.2
parso==0.8.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyMuPDF==1.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
scikit-learn==1.3.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.27.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.23.2
wcwidth==0.2.13
yarl==1.20.0
zstandard==0.23.0
